update type and move types folder location
